sistema-back/requirements.txt
asgiref==3.5.0
Django==4.0.3
django-simple-history==3.0.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
psycopg2==2.9.3
PyJWT==2.4.0
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
